requirements.txt
chromadb
duckduckgo-search
gradio
joblib
langchain
langchain-chroma
langchain-community
langchain-openai
langchain-huggingface
matplotlib
numpy
pandas
plotly
prophet
pypdf
python-dotenv
sentence-transformers
streamlit
tabulate
unstructured[pdf]